feat(backend/indicateurs): expose la trajectoire sur l'api rest
